// src/app/models/variable.ts
export class Variable {
    _id: number;
    nombre: String;
    simbolo: String;
    valor: String;
    formulaCoeficiente: String;
    isequipo: Boolean;
    elementosCalibracion: any;
    equipoId:String;
    variableEquipoId: String;
    esrepetibilidad: Boolean;
    escalibrado: Boolean;
    valoresRepetibilidad: any;
    escurvacalibracion: Boolean;
    numeroReplicas: Number;
    valoresCurvaCalibracion: any;
    puntoscalibracion: any;
    elementosIncertidumbre: any;
    incertidumbre: any;
    kcertificado:any;
    valortipodist:any;
}